[php] Refactor and Re-Organize Removed dead code, organized `AstCreator` into traits. It's clear the following still need to be handled: * `astForAnonymousClass`, `astforTraitUseStmt`, `astForAnonymousClassInstantiation` require implementations * `astForCatchStmt` needs a local variable * `astForUseStmt` and `astForGroupUseStmt` require local handling
