+++ joern-cli/frontends/php2cpg/src/main/scala/io/joern/php2cpg/astcreation/AstCreatorHelper.scala
@@ -1,0 +1,116 @@
+package io.joern.php2cpg.astcreation
+
+import io.joern.php2cpg.astcreation.AstCreator.TypeConstants
+import io.joern.php2cpg.datastructures.ArrayIndexTracker
+import io.joern.php2cpg.parser.Domain.*
+import io.joern.x2cpg.Defines.UnresolvedNamespace
+import io.joern.x2cpg.utils.AstPropertiesUtil.RootProperties
+import io.joern.x2cpg.utils.IntervalKeyPool
+import io.joern.x2cpg.{Ast, ValidationMode}
+import io.shiftleft.codepropertygraph.generated.nodes.{NewIdentifier, NewLiteral, NewNamespaceBlock}
+import io.shiftleft.semanticcpg.language.types.structure.NamespaceTraversal
+
+import java.nio.charset.StandardCharsets
+
+trait AstCreatorHelper(disableFileContent: Boolean)(implicit withSchemaValidation: ValidationMode) { this: AstCreator =>
+
+  protected val globalNamespace: NewNamespaceBlock = globalNamespaceBlock()
+  private val tmpKeyPool                           = new IntervalKeyPool(first = 0, last = Long.MaxValue)
+
+  protected def getNewTmpName(prefix: String = "tmp"): String = s"$prefix${tmpKeyPool.next.toString}"
+
+  protected def line(phpNode: PhpNode): Option[Int] = phpNode.attributes.lineNumber
+
+  protected def column(phpNode: PhpNode): Option[Int] = None
+
+  protected def lineEnd(phpNode: PhpNode): Option[Int] = None
+
+  protected def columnEnd(phpNode: PhpNode): Option[Int] = None
+
+  protected def code(phpNode: PhpNode): String = "" // Sadly, the Php AST does not carry any code fields
+
+  override protected def offset(phpNode: PhpNode): Option[(Int, Int)] = {
+    Option.when(!disableFileContent) {
+      val startPos =
+        new String(fileContent.get.getBytes.slice(0, phpNode.attributes.startFilePos), StandardCharsets.UTF_8).length
+      val endPos =
+        new String(fileContent.get.getBytes.slice(0, phpNode.attributes.endFilePos), StandardCharsets.UTF_8).length
+      (startPos, endPos)
+    }
+  }
+
+  protected def intToLiteralAst(num: Int): Ast = {
+    Ast(NewLiteral().code(num.toString).typeFullName(TypeConstants.Int))
+  }
+
+  protected def getTmpIdentifier(
+    originNode: PhpNode,
+    maybeTypeFullName: Option[String],
+    prefix: String = ""
+  ): NewIdentifier = {
+    val name         = s"$prefix${getNewTmpName()}"
+    val typeFullName = maybeTypeFullName.getOrElse(TypeConstants.Any)
+    identifierNode(originNode, name, s"$$$name", typeFullName)
+  }
+
+  protected def composeMethodFullName(methodName: String, isStatic: Boolean): String = {
+    if (methodName == NamespaceTraversal.globalNamespaceName) {
+      globalNamespace.fullName
+    } else {
+      val className       = getTypeDeclPrefix
+      val methodDelimiter = if (isStatic) StaticMethodDelimiter else InstanceMethodDelimiter
+
+      val nameWithClass = List(className, Some(methodName)).flatten.mkString(methodDelimiter)
+
+      prependNamespacePrefix(nameWithClass)
+    }
+  }
+
+  protected def prependNamespacePrefix(name: String): String = {
+    scope.getEnclosingNamespaceNames.filterNot(_ == NamespaceTraversal.globalNamespaceName) match {
+      case Nil   => name
+      case names => names.appended(name).mkString(NamespaceDelimiter)
+    }
+  }
+
+  private def getTypeDeclPrefix: Option[String] = {
+    scope.getEnclosingTypeDeclTypeName
+      .filterNot(_ == NamespaceTraversal.globalNamespaceName)
+  }
+
+  protected def codeForMethodCall(call: PhpCallExpr, targetAst: Ast, name: String): String = {
+    val callOperator = if (call.isNullSafe) "?->" else "->"
+    s"${targetAst.rootCodeOrEmpty}$callOperator$name"
+  }
+
+  protected def codeForStaticMethodCall(call: PhpCallExpr, name: String): String = {
+    val className =
+      call.target
+        .map(astForExpr)
+        .map(_.rootCode.getOrElse(UnresolvedNamespace))
+        .getOrElse(UnresolvedNamespace)
+    s"$className::$name"
+  }
+
+  protected def dimensionFromSimpleScalar(scalar: PhpSimpleScalar, idxTracker: ArrayIndexTracker): PhpExpr = {
+    val maybeIntValue = scalar match {
+      case string: PhpString =>
+        string.value
+          .drop(1)
+          .dropRight(1)
+          .toIntOption
+
+      case number => number.value.toIntOption
+    }
+
+    maybeIntValue match {
+      case Some(intValue) =>
+        idxTracker.updateValue(intValue)
+        PhpInt(intValue.toString, scalar.attributes)
+
+      case None =>
+        scalar
+    }
+  }
+
+}
